Rename SidebarContainer to CentralContainer.
