+++ new-log-viewer/src/components/CentralContainer/PanelTabs/CustomListItem.tsx
