+++ new-log-viewer/src/components/CentralContainer/PanelTabs/CustomTabPanel.tsx
