+++ new-log-viewer/src/components/CentralContainer/PanelTabs/FileInfoTab.tsx
