+++ new-log-viewer/src/components/CentralContainer/PanelTabs/TooltipTab.tsx
