+++ new-log-viewer/src/components/CentralContainer/PanelTabs/index.css
