+++ new-log-viewer/src/components/CentralContainer/ResizeHandle.tsx
