+++ new-log-viewer/src/components/CentralContainer/index.tsx
